APM_Control: factor out an AP_PitchRollController base class
